added added start to API
The AIInfo class is going to be the API, if it needs more functions they should be easy to implement. added a few things to Player class, such as remembering cards played and who they passed to and what they passed. also added a running scoreboard to the game. It only played hand by hand from what I could see